Cleanup and consolidation
tcpS.go is not main.go
